解耦API路由器和API服务器
